add module train. mv some global vars from run_opt to env
