+++ deepmd/descriptor/loc_frame.py
@@ -2,8 +2,8 @@ import numpy as np
 from typing import Tuple, List
 
 from deepmd.env import tf
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
 from deepmd.env import op_module
 from deepmd.env import default_tf_session_config
