+++ deepmd/loss/ener.py
@@ -2,8 +2,8 @@ import numpy as np
 from deepmd.env import tf
 from deepmd.common import ClassArg, add_data_requirement
 
-from deepmd.run_options import global_cvt_2_tf_float
-from deepmd.run_options import global_cvt_2_ener_float
+from deepmd.env import global_cvt_2_tf_float
+from deepmd.env import global_cvt_2_ener_float
 
 class EnerStdLoss () :
     """
