+++ deepmd/model/tensor.py
@@ -3,7 +3,7 @@ from typing import Tuple, List
 
 from deepmd.env import tf
 from deepmd.common import ClassArg
-from deepmd.run_options import global_cvt_2_ener_float, MODEL_VERSION
+from deepmd.env import global_cvt_2_ener_float, MODEL_VERSION
 from deepmd.env import op_module
 from .model_stat import make_stat_input, merge_sys_stat
