+++ deepmd/utils/neighbor_stat.py
@@ -6,7 +6,7 @@ from deepmd.env import tf
 from typing import Tuple, List
 from deepmd.env import op_module
 from deepmd.env import default_tf_session_config
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
 from deepmd.utils.data_system import DeepmdDataSystem
 
 log = logging.getLogger(__name__)
