+++ source/CMakeLists.txt
@@ -211,7 +211,7 @@ add_subdirectory (op/)
 add_subdirectory (lib/)
 if (BUILD_PY_IF)
   add_subdirectory (train/)
-  add_subdirectory (tests/)
+  # add_subdirectory (tests/)
 endif (BUILD_PY_IF)
 if (BUILD_CPP_IF) 
   add_subdirectory (api_cc/)
