+++ source/tests/common.py
@@ -3,9 +3,9 @@ import numpy as np
 import pathlib
 
 from deepmd.env import tf
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_ENER_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_ENER_FLOAT_PRECISION
 from deepmd.common import j_loader as dp_j_loader
 
 if GLOBAL_NP_FLOAT_PRECISION == np.float32 :
