+++ source/tests/test_data_modifier.py
@@ -4,12 +4,12 @@ import unittest
 from deepmd.env import tf
 
 from deepmd.common import j_must_have, data_requirement
-from deepmd.run_options import RunOptions
-from deepmd.trainer import NNPTrainer
+from deepmd.train.run_options import RunOptions
+from deepmd.train.trainer import DPTrainer
 from deepmd.utils.data_system import DeepmdDataSystem
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_ENER_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_ENER_FLOAT_PRECISION
 from deepmd.infer.ewald_recp import EwaldRecp
 from deepmd.infer.data_modifier import DipoleChargeModifier
 
@@ -50,7 +50,7 @@ class TestDataModifier (unittest.TestCase) :
         jdata = j_loader(INPUT)
 
         # init model
-        model = NNPTrainer (jdata, run_opt = run_opt)
+        model = DPTrainer (jdata, run_opt = run_opt)
         rcut = model.model.get_rcut()
 
         # init data system
