+++ source/tests/test_data_modifier_shuffle.py
@@ -5,12 +5,12 @@ import dpdata
 from deepmd.env import tf
 
 from deepmd.common import j_must_have, data_requirement
-from deepmd.run_options import RunOptions
-from deepmd.trainer import NNPTrainer
+from deepmd.train.run_options import RunOptions
+from deepmd.train.trainer import DPTrainer
 from deepmd.utils.data_system import DeepmdDataSystem
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_ENER_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_ENER_FLOAT_PRECISION
 from deepmd.infer.ewald_recp import EwaldRecp
 from deepmd.infer.data_modifier import DipoleChargeModifier
 from deepmd.infer.deep_dipole import DeepDipole
@@ -56,7 +56,7 @@ class TestDataModifier (unittest.TestCase) :
         self._setUp_data()
 
         # init model
-        model = NNPTrainer (jdata, run_opt = run_opt)
+        model = DPTrainer (jdata, run_opt = run_opt)
         rcut = model.model.get_rcut()
 
         # init data system
