+++ source/tests/test_descrpt_sea_ef_rot.py
@@ -4,9 +4,9 @@ import unittest
 
 from deepmd.env import tf
 from tensorflow.python.framework import ops
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_ENER_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_ENER_FLOAT_PRECISION
 
 from deepmd.env import op_module
 from deepmd.descriptor import DescrptSeA
