+++ source/tests/test_dipolecharge.py
@@ -6,7 +6,7 @@ from infer.convert2pb import convert_pbtxt_to_pb
 from deepmd.infer import DipoleChargeModifier
 from common import tests_path
 
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
 if GLOBAL_NP_FLOAT_PRECISION == np.float32 :
     default_places = 4
 else :
