+++ source/tests/test_embedding_net.py
@@ -7,9 +7,9 @@ from tensorflow.python.framework import ops
 
 from deepmd.utils.network import embedding_net
 
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_ENER_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_ENER_FLOAT_PRECISION
 
 class Inter(unittest.TestCase):
     def setUp (self) :
