+++ source/tests/test_ewald.py
@@ -3,9 +3,9 @@ import numpy as np
 import unittest
 from deepmd.env import tf
 
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_ENER_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_ENER_FLOAT_PRECISION
 from deepmd.infer.ewald_recp import op_module
 from deepmd.infer.ewald_recp import EwaldRecp
