+++ source/tests/test_prod_env_mat.py
@@ -5,9 +5,9 @@ import unittest
 import deepmd.op
 from deepmd.env import tf
 from deepmd.env import op_module
-from deepmd.run_options import GLOBAL_TF_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_NP_FLOAT_PRECISION
-from deepmd.run_options import GLOBAL_ENER_FLOAT_PRECISION
+from deepmd.env import GLOBAL_TF_FLOAT_PRECISION
+from deepmd.env import GLOBAL_NP_FLOAT_PRECISION
+from deepmd.env import GLOBAL_ENER_FLOAT_PRECISION
 
 class TestProdEnvMat(unittest.TestCase):
     def setUp(self):
